refactor(hub): make httpserver package compile and wire hub to it
- middleware.go becomes package httpserver: WithRecovery, WithRequestID,
  WithMetrics, CORS, statusRecorder, and the WriteErr/JSONOK/JSONError
  response helpers (error envelope unchanged).
- WithRequestID takes the base *slog.Logger and WithMetrics takes a
  MetricsRecorder interface, so httpserver has no dependency on hub.
- serverMetrics.observeRequest/webhookError become exported to satisfy
  httpserver.MetricsRecorder; webhookIntegration moves to httpserver with
  its test (it is route-pattern knowledge).
- hub keeps writeErr/jsonOK/jsonError as one-line aliases so the ~250
  existing call sites do not churn; Run composes the same middleware
  chain in the same order via the httpserver package.
- TestLogfComponentExtraction returns to pkg/hub (logging_test.go)
  because logMsg stays in hub.

Zero behavior change.
